+++ src/egon/data/datasets/power_plants/mastr_db_classes.py
@@ -1,4 +1,8 @@
+import datetime
+import json
+
 from geoalchemy2 import Geometry
+from omi.dialects import get_dialect
 from sqlalchemy import (
     Boolean,
     Column,
@@ -9,6 +13,17 @@ from sqlalchemy import (
     String,
 )
 from sqlalchemy.ext.declarative import declarative_base
+
+from egon.data import config, db
+from egon.data.metadata import (
+    context,
+    contributors,
+    generate_resource_fields_from_db_table,
+    license_dedl,
+    meta_metadata,
+    meta_metadata,
+    sources,
+)
 
 Base = declarative_base()
 
@@ -280,3 +295,154 @@ class EgonPowerPlantsStorage(Base):
 
     geometry_geocoded = Column(Boolean)
     geom = Column(Geometry("POINT", 4326), index=True, nullable=True)
+
+
+def add_metadata():
+    technologies = config.datasets()["mastr_new"]["technologies"]
+
+    target_tables = {
+        "pv": EgonPowerPlantsPv,
+        "wind": EgonPowerPlantsWind,
+        "biomass": EgonPowerPlantsBiomass,
+        "hydro": EgonPowerPlantsHydro,
+        "combustion": EgonPowerPlantsCombustion,
+        "gsgk": EgonPowerPlantsGsgk,
+        "nuclear": EgonPowerPlantsNuclear,
+        "storage": EgonPowerPlantsStorage,
+    }
+
+    deposit_id_data_bundle = config.datasets()["data-bundle"]["sources"][
+        "zenodo"
+    ]["deposit_id"]
+    deposit_id_mastr = config.datasets()["mastr_new"]["deposit_id"]
+
+    contris = contributors(["kh", "kh"])
+
+    contris[0]["date"] = "2023-03-15"
+
+    contris[0]["object"] = "metadata"
+    contris[1]["object"] = "dataset"
+
+    contris[0]["comment"] = "Add metadata to dataset."
+    contris[1]["comment"] = "Add workflow to generate dataset."
+
+    for technology in technologies:
+        target_table = target_tables[technology]
+
+        meta = {
+            "name": (
+                f"{target_table.__table_args__['schema']}."
+                f"{target_table.__tablename__}"
+            ),
+            "title": f"eGon {technology} power plants",
+            "id": "WILL_BE_SET_AT_PUBLICATION",
+            "description": (
+                f"eGon {technology} power plants status quo derived from MaStR"
+            ),
+            "language": "en-US",
+            "keywords": [technology, "mastr"],
+            "publicationDate": datetime.date.today().isoformat(),
+            "context": context(),
+            "spatial": {
+                "location": "none",
+                "extent": "Germany",
+                "resolution": "1 m",
+            },
+            "temporal": {
+                "referenceDate": (
+                    config.datasets()["mastr_new"]["egon2021_date_max"].split(
+                        " "
+                    )[0]
+                ),
+                "timeseries": {},
+            },
+            "sources": [
+                {
+                    "title": "Data bundle for egon-data",
+                    "description": (
+                        "Data bundle for egon-data: A transparent and "
+                        "reproducible data processing pipeline for energy "
+                        "system modeling"
+                    ),
+                    "path": (
+                        "https://zenodo.org/record/"
+                        f"{deposit_id_data_bundle}#.Y_dWM4CZMVM"
+                    ),
+                    "licenses": [license_dedl(attribution="© Cußmann, Ilka")],
+                },
+                {
+                    "title": (
+                        "open-MaStR power unit registry for eGo^n project"
+                    ),
+                    "description": (
+                        "Data from Marktstammdatenregister (MaStR) data using "
+                        "the data dump from 2022-11-17 for eGon-data."
+                    ),
+                    "path": (
+                        f"https://zenodo.org/record/{deposit_id_mastr}"
+                    ),
+                    "licenses": [license_dedl(attribution="© Amme, Jonathan")],
+                },
+                sources()["egon-data"],
+            ],
+            "licenses": [license_dedl(attribution="© eGon development team")],
+            "contributors": contris,
+            "resources": [
+                {
+                    "profile": "tabular-data-resource",
+                    "name": (
+                        f"{target_table.__table_args__['schema']}."
+                        f"{target_table.__tablename__}"
+                    ),
+                    "path": "None",
+                    "format": "PostgreSQL",
+                    "encoding": "UTF-8",
+                    "schema": {
+                        "fields": generate_resource_fields_from_db_table(
+                            target_table.__table_args__["schema"],
+                            target_table.__tablename__,
+                            geom_columns=["geom"],
+                        ),
+                        "primaryKey": "id",
+                    },
+                    "dialect": {"delimiter": "", "decimalSeparator": ""},
+                }
+            ],
+            "review": {"path": "", "badge": ""},
+            "metaMetadata": meta_metadata(),
+            "_comment": {
+                "metadata": (
+                    "Metadata documentation and explanation (https://github."
+                    "com/OpenEnergyPlatform/oemetadata/blob/master/metadata/"
+                    "v141/metadata_key_description.md)"
+                ),
+                "dates": (
+                    "Dates and time must follow the ISO8601 including time "
+                    "zone (YYYY-MM-DD or YYYY-MM-DDThh:mm:ss±hh)"
+                ),
+                "units": "Use a space between numbers and units (100 m)",
+                "languages": (
+                    "Languages must follow the IETF (BCP47) format (en-GB, "
+                    "en-US, de-DE)"
+                ),
+                "licenses": (
+                    "License name must follow the SPDX License List "
+                    "(https://spdx.org/licenses/)"
+                ),
+                "review": (
+                    "Following the OEP Data Review (https://github.com/"
+                    "OpenEnergyPlatform/data-preprocessing/wiki)"
+                ),
+                "none": "If not applicable use (none)",
+            },
+        }
+
+        dialect = get_dialect(meta_metadata()["metadataVersion"])()
+
+        meta = dialect.compile_and_render(dialect.parse(json.dumps(meta)))
+
+        db.submit_comment(
+            f"'{json.dumps(meta)}'",
+            target_table.__table_args__["schema"],
+            target_table.__tablename__,
+        )